fix: publish an alpine docker image
Closes #28

This supports a use-case of running custom commands inside the
typescript-tools alpine image, for example in CI.

There is an option to build different targets from a single dockerfile,
but the semantic-release plugin I'm using does not appear to support
that functionality. Instead, duplicate the original Dockerfile and
create a single alpine target.