fix: sort package json
The real reason for this commit is to publish a new version.

Looks like I don't have all the kinks worked out of this
GitHub/semantic-release workflow. I think the issue is I published a
beta version on the `beta` branch, but when I merged that into `master`
then the tag from `beta` existed on `master` and then there was no
trigger for a new release.