feat: add make-depends subcommand
This subcommand generates a GNU Makefile for a specified monorepo package.

The motivation behind this feature is to enable static, automatically
generated make targets to optimize bootstrapping in non-hoisted monorepos.

This work is heavily modeled after [Auto-Dependency Generation].

Integration with this feature will not be covered in this commit message,
but expect additional documentation shortly.

[auto-dependency generation]: http://make.mad-scientist.net/papers/advanced-auto-dependency-generation/#traditional