fix: update mtime of downloaded binary after installation
This commit adds a step to the `npm install` process, that of updating
the downloaded binary's mtime to the current time.

Apparently, the binary created by GitHub Actions retains its original
creation mtime when downloaded by the postinstall script. This interferes
with GNU Make targets that examine the mtime of the `monorepo` binary.

Instead, we accurately reflect the install time in the mtime of the
downloaded binary via the `fs.utimesSync` function offered by Node.js.