ci: use cargo-deny
Use [cargo-deny] to verify the source location from which dependencies
are obtained, and to restrict the dependencies permitted by license
used.

Looks like there is some overlap in functionality with the cargo-audit
crate, but there are no plans now to remove cargo-audit since the
maintainer-experience of the cron job run in CI is nicer (cargo-audit
will open issues on detected security vulnerabilities).

Closes #53

[cargo-deny]: https://github.com/EmbarkStudios/cargo-deny