fix: improve error messages presented to the user
This commit improves error messages presented to the user when
encountering errors parsing JSON.

This required a bit of refactoring, since we were performing too much
work up-front, and swallowing the helpful errors with a fallback (that
stops looking for a lerna manifest and starts looking for an npm 7
workspace) that happened too late.

I took this opportunity to improve the representation of package.json
manifests, which makes interaction with serde more ergonomic.